prepared teq run

+++ modular/source/3DXY.h
@@ -5,7 +5,7 @@
 #include "randStruct.h"
 namespace _3DXY {
 
-void warmup(Lattice3DXY& lat,Cluster&clust,long double beta,RandStruct& rand,int N);
+void warmup(Lattice3DXY& lat,Cluster&clust,long double beta,RandStruct& rand,long double N);
 
 void warmupMetro(Lattice3DXY& lat,long double beta,RandStruct&rand,int N);
 void wolffHistJob(long double L);
